Renaming and cleanup
