+++ heymans/routes/__init__.py
@@ -3,4 +3,4 @@ from ._replies import no_content, not_found, success, forbidden, \
 from ._quizzes import quizzes_api_blueprint
 from ._documents import documents_api_blueprint
 from ._app import app_blueprint, User
-from .google_login import google_login_blueprint
+from ._google_login import google_login_blueprint
